Angular proyect document for user

+++ AngularApp/myApp/src/app/StyleClass/styleClass-main.component.ts
@@ -1,6 +1,5 @@
 import { Component, OnInit } from '@angular/core';
 import { StyleClass } from '../models/styleClass';
-import { Style } from '../models/style';
 import { StyleClassService } from '../services/styleClass.service';
 
 @Component({
@@ -9,7 +8,7 @@ import { StyleClassService } from '../services/styleClass.service';
 })
 
 export class StyleClassMainComponent implements OnInit {
-  title = 'Style Classes';
+  title = 'Clases de estilo';
   styleClasses: Array<StyleClass>;
 
   constructor(private _styleClassService: StyleClassService) {
@@ -22,7 +21,19 @@ export class StyleClassMainComponent implements OnInit {
     )
   }
 
+  ngOnChanges(): void {
+    this._styleClassService.getStyleClasses().subscribe(
+      ((obtainedStyleClasses: Array<StyleClass>) => this.styleClasses = obtainedStyleClasses),
+      ((error: any) => console.log(error))
+    )
+    window.location.reload();
+  }
+
   deleteStyleClass(id: string): void {
-    this._styleClassService.deleteStyleClass(id).subscribe();
+    if (window.confirm('¿Esta seguro que desea eliminar el estilo?')) {
+      this._styleClassService.deleteStyleClass(id).subscribe();
+      alert("Clase de estilo eliminada");
+      window.location.reload();
+    }
   }
 }
